refactor(cli): 分离入口装配与共用宿主检查
