+++ src/eval-hosts/node/index.ts
@@ -1,4 +1,4 @@
-export * from './node-cli-composition.js';
+export * from './preflight.js';
 export * from './node-cli-evaluation-resolver.js';
 export * from './node-sample-content-resolver.js';
 export * from './safe-http-content-resolver.js';
